pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.4,<4"]

[project]
authors = [{name = "xinetzone", email = "735613050@qq.com"}]
dynamic = ["version", "description"]
license = {file = "LICENSE"}
name = "torch-book"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "d2py[dev]",
  "invoke",
  "onnx",
  "onnxscript",
  "onnxruntime",
  "tables",
  "imageio",
]

maintainers = [
  {name = "xinetzone", email = "735613050@qq.com"},
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "Framework :: Sphinx",
  "Framework :: Sphinx :: Theme",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
]

[project.urls]
Home = "https://github.com/xinetzone/torch-book"

[project.optional-dependencies]
doc = [
  "myst-nb",
  "sphinx",
  "sphinx-book-theme",
  "sphinxext-rediraffe",
  "sphinx-copybutton",
  "sphinxcontrib-mermaid",
  "sphinx-design",
  "sphinxcontrib-bibtex",
  "sphinx-thebe",
  "sphinx-comments",
  "matplotlib",
  "sphinx-automodapi",
  "sphinx-plotly-directive",
  "sphinx-sitemap",
  "sphinx-proof",
  # MMLab
  "terminaltables",
  "pycocotools",
  "shapely",
  "albumentations",
  "numba",
  "timm",
  "av",
  "click",  # required by mmagic/utils/io_utils.py
  "captum",
  "Flask-Compress",
]

coverage = [
  "pytest-cov",
  "pytest-regressions",
  "codecov",
  "torch_book[test]",
]

dev = [
  "pre-commit",
  "torch_book[coverage]",
  "controlnet_aux",
  "diffusers>=0.23.0",
  "einops",
  "face-alignment<=1.3.4",
  "facexlib",
  "lmdb",
  "lpips",
  "mediapipe",
  "numpy",
  # MMCV depends opencv-python instead of headless, thus we install opencv-python
  # Due to a bug from upstream, we skip this two version
  # https://github.com/opencv/opencv-python/issues/602
  # https://github.com/opencv/opencv/issues/21366
  # It seems to be fixed in https://github.com/opencv/opencv/pull/21382
  "opencv-python!=4.5.5.62,!=4.5.5.64",
  "pandas", # required by mmagic/models/editors/disco_diffusion/guider.py
  "Pillow",
  "resize_right",
  "tensorboard",
  "transformers>=4.27.4",
  "accelerate",
  "torch",
  "tensorflow",
  "onnx",
]
test = [
  "pytest",
  "torch_book[doc]",
]

[tool.flit.module]
name = "torch_book"

[tool.flit.sdist]
# include = ["docs/"]
exclude = ['doc', '.git', '.github']
